Organ Update part two.
This one is a bit bigger and not 100% needed, but most of the changes here make sense to me.
Lungs are better in both cold and heat, skrell are creatures that adapt to their environment so this makes sense to me.
Changes the diet to include seafood and give them water breathing, they are amphibious after all and wonderful fishers.
Their diet can be a lot of toxic stuff, so their liver should be able to deal with toxic stuff easier.

As for the most controversial trait here, Hardly wounded, I feel as though skrell having 70% more brute damage is a bit much, but fair! It makes sense for the flavor of the species that they would have skin that is easy to bruise. That being said, the skin is thick and the bones are softer with more of a rubber feel to them. Both less likely to break, but are weaker in return. A good tradeoff imo.