Adds a new solo fugitive backstory: The Invisible Man (#74415)
## About The Pull Request

A disgruntled scientist, formerly the project lead at an experimental
cloaking technology lab, now on the run for stealing the very prototypes
they helped create.


![image](https://user-images.githubusercontent.com/28870487/229234117-ef2d3951-4f57-4d62-9001-6d17808e2799.png)

You start off with a toggleable cloaking implant (makes you a bit more
invisible than a cloaked ninja), and some Stabilized Saturn-X injectors
(Which turn you COMPLETELY invisible).

The three Saturn-X injectors grant you a cumulative 30 minutes of total
invisibility (with the caveat and associated complications of you having
to be naked and not able to just take them all at once
consequence-free). It may not be as powerful as Waldo and his almighty
Knock spell, but don't underestimate how much the camouflage can do for
you in the dark.


![image](https://user-images.githubusercontent.com/28870487/229234660-7d8ded61-8997-4536-8c13-a326a1635d03.png)

So go on, blend in with the crew. When the hunters come looking, shed
your uniform and disappear...

The color of this PR is: **Purple**.
## Why It's Good For The Game

Adds some variety to solo fugitive selections (Waldo is still a
bastard). Gives the opportunity for High-End SS13 Stealth Gameplay.
## Changelog
:cl: Rhials
add: New solo fugitive backstory: The Invisible Man!
spriteadd: New purple medipen and admittedly bad camouflage icon hud
button sprite.
/:cl: