[SEMI-MODULAR] Allows observers the ability to activate pAI cards on their own (#217)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
What this PR does is that it allows observing players the ability to
join the round as a pAI without having someone in-round be needed to
activate a pAI card, wait for someone to submit a personality and then
download said personality.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->
* Closes https://github.com/Bubberstation/Bubberstation/issues/140
<!-- Please make sure to actually test your PRs. If you have not tested
your PR mention it. -->

## Testing Photographs and Procedure
<details>
<summary>Screenshots&Videos</summary>


https://user-images.githubusercontent.com/59128051/235649357-54772d46-aae6-401d-b565-066fcb92da69.mp4

</details>

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
add: Grants observers the ability to activate and possess pAI cards
without requiring someone to be on hand to activate the pAI card for
them.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Co-authored-by: nevimer <77420409+nevimer@users.noreply.github.com>