[it's almost kinda modular] sec/medhud resprites: the citadel port (#280)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

ports various hud sprites from ~~citadel~~ citadel as a port of a port,
with the original heralding from
https://github.com/shiptest-ss13/Shiptest/pull/1469 with a couple extra
touchups to make them what's essentially just the current sprites but a
little bit brighter and easier to read
also changes the sechud icons for the departmental guards to just be the
normal security officer icon but with a different departmental
background
why was the supply guard `. . !`
why was the engineering guard a `thing`
why was the medical guard a weird ass circle that looked almost
`identical` to the virologist's job icon
idk
it's much easier to tell that they're a departmental guard when they
have the security officer job icon

too lazy to post images just look at icondiffs and complain

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

<!-- Please make sure to actually test your PRs. If you have not tested
your PR mention it. -->

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
image: huds
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->