(Modular) Adds some new biogenerator products (#344)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
I was offered money by BurningCold on Discord to make code to add
several fruit juices to the biogenerator food list. I asked why it was
necessary, since some of these are in the booze or soft drink dispensers
and they told me it is because they regularly play as a prisoner, and
some plants are not allowed in the prison due to certain hazards
presented by mutations of the plants they are obtained from. Adding
these allows for ghetto bartending in the prison. Alongside these, as a
show of good will, I have also added Diethylamine and Saltpetre to the
list of chemicals, and Soy Sauce to the food list so that there's an
alternative for making it other than ordering it or having to grow
nettles, which also present their own threat to sec officers. Never let
it be said that I don't deliver.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

<!-- Please make sure to actually test your PRs. If you have not tested
your PR mention it. -->

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
code: Adds Orange Juice, Lime Juice, Lemon Juice, Berry Juice, Tomato
Juice, Grenadine Syrup, and Soy Sauce to the Food product list of the
biogenerator. Also adds Diethylamine and Saltpetre to the Chemicals
product list.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->