Let Cargo Borgs break bad (#339)
All this PR does is add a boxcutter to cargo borgs. Why? Because at the
moment they are unable to open wrapped packages at all. This is not a
massive deal on its own, however, cargoborgs are required to wrap
packages to lift and haul them. This means that they are unable to
deliver the packages that they wrap up, open them, and bring back the
empty box in one trip, just as an example of one of the reasons they
need a boxcutter.

Will this give cargoborgs a means to deal damage to others? Sure, but a
lot of borgs have the ability to do damage with one or more of their
tools far better than a boxcutter.

This was tested and I even learned that by doing this, borgs can recycle
some wrapping paper this way! yippee!

---------

Co-authored-by: projectkepler-RU <99981766+projectkepler-ru@users.noreply.github.com>
Co-authored-by: nevimer <77420409+nevimer@users.noreply.github.com>