Allows guardians and holoparas to see in the dark (#337)
## About The Pull Request
Miners are able to get holoparas. The issue? Total blindness, which
makes them not useful and literally just drags another player in with
you to do nothing due to the lack of sight.
I could split this, but I believe normal holoparas deserve this change
too.

This gives them nightvision

closes: #336

## Changelog

:cl:
fix: Holoparas can now naturally see in the dark
/:cl: