Ports Lima Station from TaleStation (#305)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Ports LimaStation from
[TaleStation](https://github.com/TaleStation/TaleStation)
Full credits to the mappers at
[TaleStation](https://github.com/TaleStation/TaleStation) &
[RussStation](https://github.com/RussStation/RussStation)

![2023-06-29 22 30
05](https://github.com/Bubberstation/Bubberstation/assets/123670072/b7386608-64fa-47b5-a2aa-9657787461c0)
![2023-06-29 22 29
58](https://github.com/Bubberstation/Bubberstation/assets/123670072/bde500a4-56dd-4694-a669-72c201f9e135)

_Taken from https://github.com/TaleStation/TaleStation/wiki/LimaStation_

**About**
LimaStation is comprised of two Z levels, the bottom (top) and the top
(bottom).

**Bottom Level**
Limas first floor comprises of never-ending maintenance tunnels.
However, the HFR room, Chemistry Lab, Xenobiology slime pens and some
service/common areas can be found down here (like botanys bee room).

**Upper Level**
Limas second floor (main floor) comprises the bulk of the departments,
as well as the general common areas (bar for instance).

Directly north is arrivals, dorms and some common areas.

Security occupies the north west portion of the station, with the chapel
right next to it.

Going east, science occupies most of the eastern wing of Lima.

South east is where the Bridge, and further out, AI sat/Tcomms is
located.

South west is atmospherics, with directly west being the bulk of
engineering and the engine room itself.

In the center of the station is the bulk of service on the left, with
medical on the right.

Cargo is south of the bar, but north of departures.

**LimaStation History**
LimaStation is originally from
[RussStation](https://github.com/RussStation/RussStation), but was
ported by [SQNZTB](https://github.com/sqnztb) on December 8th, 2020 to
[TaleStation](https://github.com/TaleStation/JollyStation-Commit-Archive/commit/331c2f04cb426ae60922037c0ec249819792ae1d).
It was merged into code on January 28th, 2021.

Since the initial port, LimaStation has gone under many drastic changes
since it was brought to the (formerly) JollyStation. LimaStation was
original one Z level, however, a
[rework](https://github.com/TaleStation/TaleStation/commit/cd2ac245b7d025283e883cb3bcc7c6370d2652a6)
on August 11th, 2021 by [SQNZTB](https://github.com/sqnztb) revived Lima
to be what we know it as today, making it Multi-Z.

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
add: Lima Station map, ported from TaleStation. A modest map for modest
people.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->