Brings back Security Medic (#268)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Security Medic was removed in the latest upstream merge. This undoes
that, [as requested by the person who PRed said
merge](https://github.com/Bubberstation/Bubberstation/pull/236). I've
tested that the role can spawn, and that all sprites function.

## Changelog

:cl:
fix: un-deleted the Security Medic
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->

---------

Co-authored-by: Cyprex <35031555+Cyprex@users.noreply.github.com>
Co-authored-by: nevimer <77420409+nevimer@users.noreply.github.com>