removes BUG that causes DAMAGE to ANTAG (#403)
so errrrm the HoS revolver had its internal mag set wrong
so it just spawned with a .357 mag instead
which with the 1.6x proj. multiplier meant it was just an objectively
more powerful .357. For free. Roundstart guaranteed.
 oh and couldn't be reloaded with the ammo you got supplied with
Ruh roh!


![image](https://github.com/Bubberstation/Bubberstation/assets/8881105/7bf67e29-1bfc-49d6-bd13-227cb58bb1c2)

:cl:
fix: The HoS revolver can now be reloaded
fix: The HoS revolver is no longer an objectively better .357
/:cl: