shrink size of breasts_double on latex catsuit (#352)
## About The Pull Request

the latex catsuit has a cool little feature on this server, where
depending on what kind of breasts your char has (double, quadruple,
sextuple) it adds another sprite ontop of them. thing is that the size
of the sprite for `breasts_double` is comically bigger than what's
normally on the suit. I went ahead and edited the sprite so that its not
ballooning off your player.

<!-- Please make sure to actually test your PRs. If you have not tested
your PR mention it. -->

## Changelog

:cl:
image: Modified breasts_double for both the item and mob sprite of the
latex catsuit
/:cl:


![image](https://github.com/Bubberstation/Bubberstation/assets/56202842/5c04a0a8-00d0-43a1-9785-a57a4ccecaeb)


![image](https://github.com/Bubberstation/Bubberstation/assets/56202842/6788defe-e15a-4d87-9f09-517830258c61)
loaded my own game and viewed new sprite. perfection.

---------

Co-authored-by: nevimer <77420409+nevimer@users.noreply.github.com>