[Modular] Adds the Tactical  Maid Outfit for  Donors.  (#441)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request
This time with revengence.

Adds the full Syndicate maid outfit to the loadout but without its stats
bonus and syndicate part replaced with tactical.
Suggested changes are welcome and I will do them as soon as possible. 
This item was requested by skyefree. 

This was in another PR, but I think I broke the actual rebasing and
wasn't sure if I was doing it right. So I closed that PR and made this
one since It was much easier to just make a new branch based on the up
to date master and just quickly put the new files in.

Original PR was  https://github.com/Bubberstation/Bubberstation/pull/307

Sorry about the hassle. I'm not that experienced with using Git.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

<!-- Please make sure to actually test your PRs. If you have not tested
your PR mention it. -->

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
add: Adds Donor Items: Tactical Maid Outfit/Sleeves/Headband.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->