Restores Low end Space sprites (#442)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

tgstation made (awful) changes recently that made space look like THIS

![dreamseeker_YuYFvv77q1](https://user-images.githubusercontent.com/77420409/208272381-2b91b4f1-04e5-4c8c-bd4c-225b6386f575.png)

Then like THIS (not live yet on skyrat).

![dreamseeker_AgEEkhRDkR](https://user-images.githubusercontent.com/77420409/208272386-1a886d28-ee54-4487-b987-62095d86ddcd.png)

For what? shaving a whole fraction of a second off, hypothetically, of
init time? The job of the server is to pre-bake things as client-side
rendering is pretty much non-existent beyond filters and other
particular things. Not everyone likes the parralax, feels comfortable
with the parallax or even otherwise can run it on their hardware at
cost. We don't need to shave a fraction of a second off init because
it's pretty damn fast anyways, and, once I am done BENCHMARKING on a
CELERON TABLET I will prove it silly. I personally think init time saves
like this are fetishized when it throws the user experience out of the
fucking window, especially for broke/third world gamers who run systems
with like, 2GB of memory. We have players running truly destitute
systems.
## Before

![dreamseeker_vT1uJeDmgx](https://user-images.githubusercontent.com/77420409/208272512-628559ac-8c2a-4f29-9863-a998257c890f.png)
## After

![dreamseeker_lBhaey7ZFd](https://user-images.githubusercontent.com/77420409/208272521-4007b674-7126-4599-b786-fbfc512995bb.png)

I could brand this as saving 4 seconds of init time, to be deceiving.
This is with 64GB of ram on a Ryzen 5800H, Samsung 980 EVO NVME with
lowmememorymode defined.
## How This Contributes To The Skyrat Roleplay Experience

This readds a decent looking space option for those seeking every ounce
of performance in their client in exchange for the server using a
fraction of a second more CPU during init.


## Proof of Testing

<!-- Include any screenshots/videos/debugging steps of the code
functioning successfully, between the </summary> and </details> code
blocks. -->

<details>
<summary>Screenshots/Videos</summary>
  

![dreamseeker_lBhaey7ZFd](https://user-images.githubusercontent.com/77420409/208272553-53fc3f89-e4bc-4cc9-bf20-97356670c861.png)

</details>

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
add: Old space returns for those who choose it. 
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->