Jukeboksss Update (#462)
## About The Pull Request

Purchasable jukebox for 10k credits

WARNING: YOU WILL NEED BAR ACCESS

## Why It's Good For The Game

We added a shit ton of songs into it only for it to be admin only. There
is no reason for it to not be accessible to crew
(And yes I know it spawns in ruins and on birdshot. That isn't easily
accessible)

## Changelog

:cl:
add: Jukeboxes can now be purchased via the cargo shuttle
/:cl:

---------

Co-authored-by: nevimer <77420409+nevimer@users.noreply.github.com>