Reebe touchups (#454)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

This change updates Reebe, making it easier for a clockwork cultist to
initially defend as well as adding tools to help them get started with
re-fortifying it.

## Why It's Good For The Game

In order for a clockwork cultist to be able to get to reebe they must
sacrifice a blood cultist. However, in a round with a blood cult the
blood cultists are easily able to teleport to Reebe thanks to the
teleport runes that are there by default. That, along with the lack of
any form of cover for what is typically a solo clockwork cultist vs the
entire blood cult is crippling as it means the entire blood cult can
actually get to reebe before you as the solo clockwork cultist. This is
so bad because the clockwork cult revolves around being on the
defensive, almost all of your sigils and structures revolve around
having a defender's advantage.

This PR makes it so that there is cover for the clockwork cultist from
lasers and sight as well as a single of each structure, thus making it
easier for them to use their natural defenders advantage to offset the
potential attack from the blood cult.
There is also added some bronze and a replica fabricator to the armory
section of Reebe, making it easier for a clockwork cultist to begin
reinforcing and rebuilding Reebe.

## Changelog

:cl:
add: Adds a Replica Fabricator and some sheets of bronze to fill it to
the Reebe armory
add: Adds a set of doors to the Armory of Reebe
balance: Changes Reebe's central room doors to be opaque
balance: Changes Reebe's central room window's to be walls
balance: Removes one of the 2 teleport runes at Reebe
/:cl:

Co-authored-by: nevimer <77420409+nevimer@users.noreply.github.com>