Re-privates my dono item (#483)
<## About The Pull Request
I did my dono item I was going to do for Bubbers upstream instead, to
save making two separate PRs for the same item. Plus, no need to
modularise it. It was said that private items were allowed, and this one
is pretty exclusively meaningless to anyone else. It got understandably
un-privatised during an upstream merge.

## Why It's Good For The Game
I guess it makes one person more likely to play, and +1 person is good
for the game?

## Changelog
It's kinda only player-facing for one player.