Makes plushies huggable to reduce stress from monophobia (#507)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Monophobia will now check for your hands for your pets/plushies too.
Rebalances the cost to be more in pair with RDS, given it not only has
hallucinations symptoms but can also kill you instantly if you are
unlucky.
Makes you able to hug plushies for companion, trading off having to have
a full hand for it.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

<!-- Please make sure to actually test your PRs. If you have not tested
your PR mention it. -->

## Why It's Good For The Game

Adds a new way to play with the monophobia quirk, making it more
appealing to be used. Also adds another use for plushies, as they don't
usually do much other than make noise.

You will be able to hold plushies for comfort if you have monophobia,
reducing your stress. This makes sure you can play the game without
randomly dying to a heart attack and encourages the quirk to be
roleplayed in a unique way too.

It's more fun than having to rely on a random cargo pet that you took a
while to get just to not really care about it and only use it to not
feel alone. Plushies with unique names are usually not replaceable, so
that's more fun having to care about something you can put at least a
bit of lore into!

The plushies are a cheaper alternative other than carrying pets around.
They are flammable, can easily be stolen and any sharp tool will rip out
their stuffing, making them not comfortable to hug anymore. You can't
just toss the plushie in a corner, you do need it close to you for any
stress to be reduced.

Also gives you a nice excuse to be carrying a big shark plushie 24/7.

![dreamseeker_j3rDCTCW12](https://github.com/Bubberstation/Bubberstation/assets/132854285/36676a53-e649-40ee-bdeb-36b9159b0784)

![dreamseeker_Y3S9PHfLHI](https://github.com/Bubberstation/Bubberstation/assets/132854285/604744c2-cccc-4c14-a71c-02ad9333aa9a)

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
add: You can now hug plushies to reduce stress if you have monophobia.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->