Remove Haloperidol from Borgs (#506)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- You can view Contributing.MD for a detailed description of the pull
request process. -->

## About The Pull Request

Remove Haloperidol, a sleep chemical from borgs. Haloperidol is a busted
chemical for combat and it has to go. This is a bandaid fix to the
instant injection issue, but until I can convince more people, this'll
have to do. I will be experimenting with the injection timer. I removed
Haloperidol from that list. This won't break.


![image](https://github.com/Bubberstation/Bubberstation/assets/32783144/8e4faf78-98a7-4406-886f-e491ffa0d8e4)


<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review and even discourage maintainers from
merging your PR! -->

<!-- Please make sure to actually test your PRs. If you have not tested
your PR mention it. -->

## Why It's Good For The Game

Haloperidol is a busted chemical which puts the victim to sleep,
allowing any borg to click-combat insta-inject with modsuit penetration
any individual they want, taking them out of the fight in which they can
then pump them full of more. Sorry but no.

<!-- Argue for the merits of your changes and how they benefit the game,
especially if they are controversial and/or far reaching. If you can't
actually explain WHY what you are doing will improve the game, then it
probably isn't good for the game in the first place. -->

## Changelog

<!-- If your PR modifies aspects of the game that can be concretely
observed by players or admins you should add a changelog. If your change
does NOT meet this description, remove this section. Be sure to properly
mark your PRs to prevent unnecessary GBP loss. You can read up on GBP
and it's effects on PRs in the tgstation guides for contributors. Please
note that maintainers freely reserve the right to remove and add tags
should they deem it appropriate. You can attempt to finagle the system
all you want, but it's best to shoot for clear communication right off
the bat. -->

:cl:
del: Removed Haloperidol from Medical Borg Hyposprays
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->