Crew monitor sorting overhaul (#521)
## About The Pull Request


![image](https://github.com/Bubberstation/Bubberstation/assets/8602857/4f0d4aa1-f3ef-4d7b-ace4-0fe709f35f67)

Adds a new icon for near crit.
Heads are sorted at the top of the manifest.
(Not pictured) Heads are now blue instead of poop gold.
Prisoners and assistants are given their own new colors.
(Not pictured) Unknowns are placed below assistants, above prisoners.
Central Command roles are placed above captain.
Other roles (mostly off-station) are placed below prisoners.
A crewmember with 50 or more damage is placed at the top, regardless of
rank. The higher the damage, the higher they are placed.

## Why It's Good For The Game

This willl help medical players (and the AI) no get eye strain when
scrolling through the list of 60 crewmembers to see which one is fucked.

## Changelog

:cl: BurgerBB
qol: Improves the crew monitor sorting and design so it's less annoying
to use.
/:cl:

<!-- Both :cl:'s are required for the changelog to work! You can put
your name to the right of the first :cl: if you want to overwrite your
GitHub username as author ingame. -->
<!-- You can use multiple of the same prefix (they're only used for the
icon ingame) and delete the unneeded ones. Despite some of the tags,
changelogs should generally represent how a player might be affected by
the changes rather than a summary of the PR's contents. -->